WIP IVisitor Cleanup
