+++ Code/Interpreter/Interpreter/Ast/Nodes/ExpressionNodes/Expressions/FactorOperation.cs
@@ -1,6 +1,6 @@
 ﻿using Interpreter.Ast.Visitors;
 
-namespace Interpreter.Ast.Nodes.ExpressionNodes
+namespace Interpreter.Ast.Nodes.ExpressionNodes.Expressions
 {
     public sealed class FactorOperation : OperationNode
     {
