+++ Code/Interpreter/Interpreter/Ast/Nodes/ExpressionNodes/Expressions/SumOperation.cs
@@ -1,11 +1,9 @@
 ﻿using Interpreter.Ast.Visitors;
 
-namespace Interpreter.Ast.Nodes.ExpressionNodes
+namespace Interpreter.Ast.Nodes.ExpressionNodes.Expressions
 {
-    public sealed class TerminalExpression : ExpressionNode
+    public sealed class SumOperation : OperationNode
     {
-        public ExpressionNode Child { get; set; }
-        
         public override void Accept(IExpressionVisitor visitor)
         {
             visitor.Visit(this);
