+++ Code/Interpreter/Interpreter/Ast/Nodes/ExpressionNodes/Expressions/TerminalExpression.cs
@@ -1,0 +1,14 @@
+﻿using Interpreter.Ast.Visitors;
+
+namespace Interpreter.Ast.Nodes.ExpressionNodes.Expressions
+{
+    public sealed class TerminalExpression : ExpressionNode
+    {
+        public ExpressionNode Child { get; set; }
+        
+        public override void Accept(IExpressionVisitor visitor)
+        {
+            visitor.Visit(this);
+        }
+    }
+}
