+++ Code/Interpreter/Interpreter/Ast/Nodes/ExpressionNodes/Values/ArrayNode.cs
@@ -1,13 +1,13 @@
 ﻿using System.Collections.Generic;
 using Interpreter.Ast.Visitors;
 
-namespace Interpreter.Ast.Nodes.ValueNodes
+namespace Interpreter.Ast.Nodes.ExpressionNodes.Values
 {
     public sealed class ArrayNode : ValueNode
     {
         public List<ValueNode> Values { get; set; }
 
-        public override void Accept(IValueVisitor visitor)
+        public override void Accept(IExpressionVisitor visitor)
         {
             visitor.Visit(this);
         }
