+++ Code/Interpreter/Interpreter/Ast/Nodes/ExpressionNodes/Values/FloatValue.cs
@@ -1,12 +1,12 @@
 ﻿using Interpreter.Ast.Visitors;
 
-namespace Interpreter.Ast.Nodes.ValueNodes
+namespace Interpreter.Ast.Nodes.ExpressionNodes.Values
 {
     public sealed class FloatValue : ValueNode
     {
         public float Value { get; set; }
 
-        public override void Accept(IValueVisitor visitor)
+        public override void Accept(IExpressionVisitor visitor)
         {
             visitor.Visit(this);
         }
