+++ Code/Interpreter/Interpreter/Ast/Nodes/ExpressionNodes/Values/IdentifierValue.cs
@@ -1,9 +1,11 @@
 ﻿using Interpreter.Ast.Visitors;
 
-namespace Interpreter.Ast.Nodes.ExpressionNodes
+namespace Interpreter.Ast.Nodes.ExpressionNodes.Values
 {
-    public sealed class SumOperation : OperationNode
+    public sealed class IdentifierValue : ValueNode
     {
+        public string Value { get; set; }
+
         public override void Accept(IExpressionVisitor visitor)
         {
             visitor.Visit(this);
