+++ Code/Interpreter/Interpreter/Ast/Nodes/ExpressionNodes/Values/IntValue.cs
@@ -1,11 +1,11 @@
 ﻿using Interpreter.Ast.Visitors;
 
-namespace Interpreter.Ast.Nodes.ValueNodes
+namespace Interpreter.Ast.Nodes.ExpressionNodes.Values
 {
     public sealed class IntValue : ValueNode
     {
         public int Value { get; set; }
-        public override void Accept(IValueVisitor visitor)
+        public override void Accept(IExpressionVisitor visitor)
         {
             visitor.Visit(this);
         }
