+++ Code/Interpreter/Interpreter/Ast/Nodes/ExpressionNodes/Values/MemberAccess.cs
@@ -1,13 +1,13 @@
 ﻿using System.Collections.Generic;
 using Interpreter.Ast.Visitors;
 
-namespace Interpreter.Ast.Nodes.ValueNodes
+namespace Interpreter.Ast.Nodes.ExpressionNodes.Values
 {
     public sealed class MemberAccess : ValueNode
     {
         public readonly List<string> Identifiers = new();
 
-        public override void Accept(IValueVisitor visitor)
+        public override void Accept(IExpressionVisitor visitor)
         {
             visitor.Visit(this);
         }
