+++ Code/Interpreter/Interpreter/Ast/Nodes/ExpressionNodes/Values/ValueNode.cs
@@ -1,0 +1,6 @@
+﻿namespace Interpreter.Ast.Nodes.ExpressionNodes.Values
+{
+    public abstract class ValueNode : ExpressionNode
+    {
+    }
+}
