--- Code/Interpreter/Interpreter/Ast/Nodes/ValueNodes/IdentifierValue.cs
@@ -1,14 +1,0 @@
-﻿using Interpreter.Ast.Visitors;
-
-namespace Interpreter.Ast.Nodes.ValueNodes
-{
-    public sealed class IdentifierValue : ValueNode
-    {
-        public string Value { get; set; }
-
-        public override void Accept(IValueVisitor visitor)
-        {
-            visitor.Visit(this);
-        }
-    }
-}
