--- Code/Interpreter/Interpreter/Ast/Nodes/ValueNodes/ValueNode.cs
@@ -1,8 +1,0 @@
-﻿using Interpreter.Ast.Visitors;
-
-namespace Interpreter.Ast.Nodes.ValueNodes
-{
-    public abstract class ValueNode : Node<IValueVisitor>
-    {
-    }
-}
